Move spooky test that doesn't work unless in top level test file to top level test file
